refactor(group-quiz): extract in-progress quiz view (verbatim) [2/2]
Fold the two contiguous `session.status === "in_progress"` blocks
(question card + nav + post-submit waiting/reveal, and the rare
questions-loading state) verbatim into
_components/group-quiz-quiz-view.tsx; the component picks the same
sub-block the page's questions.length conditionals did, so render output
is identical by construction. Owns its own GroupQuiz.session i18n.
tsc + build + 93 tests green. page.tsx 861->723 LOC (1637->723, -56%
across the sweep). Verbatim-safe parts done; remaining interwoven
lobby/shared-chrome left as-is per scope.
